refactor socket provider, clean up console logs, clean up unused imports
