+++ client/components/PlayerList.js
@@ -1,25 +1,10 @@
-import axios from 'axios';
 import React, { useState, useEffect, useContext } from 'react';
 import 'materialize-css';
-import {
-  Container,
-  Button,
-  TextInput,
-  Dropdown,
-  Divider,
-  Icon,
-  Card,
-  Row,
-  Col,
-  Select,
-} from 'react-materialize';
-import Cycle from './Cycle';
 import styles from './css/Game.module.css';
-import { SocketContext } from '../components/context/socket';
+import { SocketContext } from './context/socket';
 import { UserContext } from './context/user';
 
 const PlayerList = () => {
-  // // socket connection logic
   const socket = useContext(SocketContext);
   const [user, _] = useContext(UserContext);
